+++ smok/client/client.py
@@ -309,10 +309,8 @@ class SMOKDevice(Closeable, metaclass=ABCMeta):
         if not delayed_boot:
             if self.use_ngtt:
                 self.sync_worker = NGTTSyncWorker(self)
-                logger.debug('Using NGTT API to sync data and logs')
             else:
                 self.sync_worker = HTTPSyncWorker(self)
-                logger.debug('Using HTTP API to sync data and logs')
             if not dont_obtain_orders or not dont_do_predicates or not dont_do_pathpoints or \
                     not dont_do_baobs:
                 self.executor.start()